カスタムhookを追加
